+++ backend/spring-routie/src/main/java/routie/business/like/ui/v1/PlaceLikeControllerV1.java
@@ -1,4 +1,4 @@
-package routie.business.placelike.ui.v1;
+package routie.business.like.ui.v1;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import routie.business.placelike.application.PlaceLikeService;
+import routie.business.like.application.PlaceLikeService;
 
 @RestController
 @RequiredArgsConstructor
