+++ backend/spring-routie/src/test/java/routie/business/place/ui/v2/PlaceControllerV2Test.java
@@ -19,9 +19,9 @@ import routie.business.place.domain.PlaceBuilder;
 import routie.business.place.domain.PlaceRepository;
 import routie.business.place.ui.dto.response.PlaceListResponseV2;
 import routie.business.place.ui.dto.response.PlaceListResponseV2.PlaceCardResponseV2;
-import routie.business.placelike.domain.PlaceLike;
-import routie.business.placelike.domain.PlaceLikeBuilder;
-import routie.business.placelike.domain.PlaceLikeRepository;
+import routie.business.like.domain.PlaceLike;
+import routie.business.like.domain.PlaceLikeBuilder;
+import routie.business.like.domain.PlaceLikeRepository;
 import routie.business.routiespace.domain.RoutieSpace;
 import routie.business.routiespace.domain.RoutieSpaceIdentifierProvider;
 import routie.business.routiespace.domain.RoutieSpaceRepository;
